requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
autoflake==2.2.1
bcrypt==4.0.1
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-cache2==0.2.1
fastapi-versioning==0.10.0
flake8==6.1.0
frozenlist==1.4.0
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
install==1.3.5
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
pendulum==2.1.2
platformdirs==4.0.0
pluggy==1.3.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
pyasn1==0.5.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyflakes==3.1.0
pyright==1.1.336
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
